test(ai): split planning_helpers_test under non-comment line gate (Refs #3749)
Move the diplomatic-cooldown and at-war peace/order scoring helper groups
into a new planning_helpers_part3_test.dart so the original file drops below
the repo.dart_file_non_comment_line_size 1000-line limit (was 1030).
